Report tree node meta inconsistencies to visitor
Also add check for misaligned allocOffset and allocSpace overlapping
active keys.

+++ community/index/src/main/java/org/neo4j/index/internal/gbptree/GBPTree.java
@@ -1213,7 +1213,7 @@ public class GBPTree<KEY,VALUE> implements Closeable
      * Print node with given id to System.out, if node with id exists.
      * @param id the page id of node to print
      */
-    void printNode( int id ) throws IOException
+    void printNode( long id ) throws IOException
     {
         if ( id < freeList.lastId() )
         {
